Remove the CLI approach
